+++ source/renderer/util/editor-tabs.js
@@ -215,6 +215,7 @@ module.exports = class EditorTabs {
   _makeElement (file, active = false, clean = true, transient = false) {
     // First determine the display title (either filename or frontmatter title)
     let displayTitle = file.name
+    if (file.firstHeading && global.config.get('display.useFirstHeadings')) displayTitle = file.firstHeading
     if (file.frontmatter && file.frontmatter.title) displayTitle = file.frontmatter.title
 
     // Then create the document div